Introducing command line arguments support for all psl interfaces.
The psl-examples and any program written with psl takes only command line arguments through maven infra(for groovy and java) or from run scripts(python).
This changelist supports having command line arguments to run scripts of all interfaces. Currently this code has been tested only with simple acquaintances. Once changes for this is in place, will try on other examples and then raise a PR on psl-examples repo.
