+++ psl-cli/src/main/java/org/linqs/psl/cli/Launcher.java
@@ -22,7 +22,6 @@ import org.linqs.psl.application.inference.InferenceApplication;
 import org.linqs.psl.application.inference.MPEInference;
 import org.linqs.psl.application.learning.weight.WeightLearningApplication;
 import org.linqs.psl.application.learning.weight.maxlikelihood.MaxLikelihoodMPE;
-import org.linqs.psl.config.Config;
 import org.linqs.psl.database.DataStore;
 import org.linqs.psl.database.Database;
 import org.linqs.psl.database.Partition;
@@ -40,26 +39,14 @@ import org.linqs.psl.model.rule.UnweightedGroundRule;
 import org.linqs.psl.model.rule.WeightedGroundRule;
 import org.linqs.psl.model.term.Constant;
 import org.linqs.psl.parser.ModelLoader;
+import org.linqs.psl.parser.CommandLineLoader;
 import org.linqs.psl.util.Reflection;
 import org.linqs.psl.util.StringUtils;
 import org.linqs.psl.util.Version;
 
 import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.CommandLineParser;
-import org.apache.commons.cli.DefaultParser;
-import org.apache.commons.cli.HelpFormatter;
-import org.apache.commons.cli.Option;
-import org.apache.commons.cli.OptionGroup;
-import org.apache.commons.cli.Options;
-import org.apache.commons.cli.ParseException;
 import org.apache.commons.configuration2.ex.ConfigurationException;
-import org.apache.log4j.BasicConfigurator;
-import org.apache.log4j.ConsoleAppender;
-import org.apache.log4j.PatternLayout;
-import org.apache.log4j.Priority;
-import org.apache.log4j.PropertyConfigurator;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -72,10 +59,9 @@ import java.net.InetAddress;
 import java.net.URL;
 import java.net.UnknownHostException;
 import java.nio.file.Paths;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
+
 import java.util.Set;
 import java.util.regex.Pattern;
 
@@ -84,33 +70,27 @@ import java.util.regex.Pattern;
  * Supports inference and supervised parameter learning.
  */
 public class Launcher {
-    // Command line options.
-    public static final String OPTION_HELP = "h";
-    public static final String OPTION_HELP_LONG = "help";
     public static final String OPERATION_INFER = "i";
     public static final String OPERATION_INFER_LONG = "infer";
+    public static final String OPTION_HELP = "h";
+    public static final String OPTION_HELP_LONG = "help";
     public static final String OPERATION_LEARN = "l";
     public static final String OPERATION_LEARN_LONG = "learn";
-
-    public static final String OPTION_DATA = "d";
-    public static final String OPTION_DATA_LONG = "data";
+    
     public static final String OPTION_DB_H2_PATH = "h2path";
     public static final String OPTION_DB_POSTGRESQL_NAME = "postgres";
+    public static final String OPTION_OUTPUT_GROUND_RULES_LONG = "groundrules";
+    public static final String OPTION_DATA = "d";
+    public static final String OPTION_DATA_LONG = "data";
     public static final String OPTION_EVAL = "e";
     public static final String OPTION_EVAL_LONG = "eval";
     public static final String OPTION_INT_IDS = "int";
-    public static final String OPTION_INT_IDS_LONG = "int-ids";
-    public static final String OPTION_LOG4J = "4j";
-    public static final String OPTION_LOG4J_LONG = "log4j";
     public static final String OPTION_MODEL = "m";
     public static final String OPTION_MODEL_LONG = "model";
     public static final String OPTION_OUTPUT_DIR = "o";
     public static final String OPTION_OUTPUT_DIR_LONG = "output";
-    public static final String OPTION_OUTPUT_GROUND_RULES_LONG = "groundrules";
     public static final String OPTION_OUTPUT_SATISFACTION_LONG = "satisfaction";
     public static final String OPTION_PROPERTIES = "D";
-    public static final String OPTION_PROPERTIES_FILE = "p";
-    public static final String OPTION_PROPERTIES_FILE_LONG = "properties";
     public static final String OPTION_VERSION = "v";
     public static final String OPTION_VERSION_LONG = "version";
 
@@ -130,89 +110,9 @@ public class Launcher {
     private CommandLine options;
     private Logger log;
 
-    private Launcher(CommandLine options) {
-        this.options = options;
-        this.log = initLogger();
-        initConfig();
-    }
-
-    /**
-     * Initializes logging.
-     */
-    private Logger initLogger() {
-        Properties props = new Properties();
-
-        if (options.hasOption(OPTION_LOG4J)) {
-            try {
-                props.load(new FileReader(options.getOptionValue(OPTION_LOG4J)));
-            } catch (IOException ex) {
-                throw new RuntimeException("Failed to read logger configuration from a file.", ex);
-            }
-        } else {
-            // Setup a default logger.
-            props.setProperty("log4j.rootLogger", "INFO, A1");
-            props.setProperty("log4j.appender.A1", "org.apache.log4j.ConsoleAppender");
-            props.setProperty("log4j.appender.A1.layout", "org.apache.log4j.PatternLayout");
-            props.setProperty("log4j.appender.A1.layout.ConversionPattern", "%-4r [%t] %-5p %c %x - %m%n");
-        }
-
-        // Load any options specified directly on the command line (override standing options).
-        for (Map.Entry<Object, Object> entry : options.getOptionProperties("D").entrySet()) {
-            String key = entry.getKey().toString();
-
-            if (!key.startsWith("log4j.")) {
-                continue;
-            }
-
-            props.setProperty(key, entry.getValue().toString());
-        }
-
-        // Log4j is pretty picky about it's thresholds, so we will specially set one option.
-        if (props.containsKey("log4j.threshold")) {
-            props.setProperty("log4j.rootLogger", props.getProperty("log4j.threshold") + ", A1");
-        }
-
-        // Some deps use java.util.logging, so we will silence their loggers.
-        java.util.logging.Logger rootJavaLogger = java.util.logging.LogManager.getLogManager().getLogger("");
-        rootJavaLogger.setLevel(java.util.logging.Level.SEVERE);
-        for (java.util.logging.Handler handler : rootJavaLogger.getHandlers()) {
-            handler.setLevel(java.util.logging.Level.SEVERE);
-        }
-
-        PropertyConfigurator.configure(props);
-        return LoggerFactory.getLogger(Launcher.class);
-    }
-
-    /**
-     * Initialize log4j with a default logger.
-     * Only to be used with short CLI runs: --version or --help.
-     */
-    private static void initDefaultLogger() {
-        Properties props = new Properties();
-
-        props.setProperty("log4j.rootLogger", "INFO, A1");
-        props.setProperty("log4j.appender.A1", "org.apache.log4j.ConsoleAppender");
-        props.setProperty("log4j.appender.A1.layout", "org.apache.log4j.PatternLayout");
-        props.setProperty("log4j.appender.A1.layout.ConversionPattern", "%-4r [%t] %-5p %c %x - %m%n");
-
-        PropertyConfigurator.configure(props);
-    }
-
-    /**
-     * Loads configuration.
-     */
-    private void initConfig() {
-        // Load a properties file that was specified on the command line.
-        if (options.hasOption(OPTION_PROPERTIES_FILE)) {
-            String propertiesPath = options.getOptionValue(OPTION_PROPERTIES_FILE);
-            Config.loadResource(propertiesPath);
-        }
-
-        // Load any options specified directly on the command line (override standing options).
-        for (Map.Entry<Object, Object> entry : options.getOptionProperties("D").entrySet()) {
-            String key = entry.getKey().toString();
-            Config.setProperty(key, entry.getValue());
-        }
+    private Launcher(CommandLineLoader commandLineLoader) {
+        this.options = commandLineLoader.options;
+        this.log = commandLineLoader.log;
     }
 
     /**
@@ -534,6 +434,36 @@ public class Launcher {
         dataStore.close();
     }
 
+    private static boolean isInputSanitized(CommandLineLoader commandLineLoader) {
+        CommandLine options = commandLineLoader.options;
+        // if only help or version was queried return
+        if ((options.hasOption(CommandLineLoader.OPTION_HELP)) || 
+        (options.hasOption(CommandLineLoader.OPTION_VERSION))) {
+            return false;
+        }
+
+        // Data and model are required.
+        // (We don't enforce them earlier so we can have successful runs with help and version.)
+
+        if (!(options.hasOption(OPTION_DATA))) {
+            System.out.println(String.format("Missing required option: --%s/-%s.", OPTION_DATA_LONG, OPTION_DATA));
+            //shrbs:FIXME getHelpFormatter().printHelp("psl", options, true);
+            return false;
+        }
+        if (!(options.hasOption(OPTION_MODEL))) {
+            System.out.println(String.format("Missing required option: --%s/-%s.", OPTION_MODEL_LONG, OPTION_MODEL));
+            //shrbs:FIXME getHelpFormatter().printHelp("psl", options, true);
+            return false;
+        }
+        //FIXME: learn and infer prints
+        if (!(options.hasOption(OPERATION_INFER)) && (!(options.hasOption(OPERATION_LEARN)))) {
+            System.out.println(String.format("Missing required option: --%s/-%s.", OPERATION_INFER_LONG, OPERATION_INFER));
+            //shrbs:FIXME getHelpFormatter().printHelp("psl", options, true);
+            return false;
+        }
+        return true;
+    }
+
     private static String getHostname() {
         String hostname = "unknown";
 
@@ -546,263 +476,17 @@ public class Launcher {
         return hostname;
     }
 
-    private static Options setupOptions() {
-        Options options = new Options();
-
-        OptionGroup mainCommand = new OptionGroup();
-
-        mainCommand.addOption(Option.builder(OPERATION_INFER)
-                .longOpt(OPERATION_INFER_LONG)
-                .desc("Run MAP inference." +
-                        " You can optionally supply a fully qualified name for an inference application" +
-                        " (defaults to " + DEFAULT_IA + ").")
-                .hasArg()
-                .argName("inferenceMethod")
-                .optionalArg(true)
-                .build());
-
-        mainCommand.addOption(Option.builder(OPERATION_LEARN)
-                .longOpt(OPERATION_LEARN_LONG)
-                .desc("Run weight learning." +
-                        " You can optionally supply a fully qualified name for a weight learner" +
-                        " (defaults to " + DEFAULT_WLA + ").")
-                .hasArg()
-                .argName("learner")
-                .optionalArg(true)
-                .build());
-
-        // Make sure that help and version are in the main group so a successful run can use them.
-
-        mainCommand.addOption(Option.builder(OPTION_HELP)
-                .longOpt(OPTION_HELP_LONG)
-                .desc("Print this help message and exit")
-                .build());
-
-        mainCommand.addOption(Option.builder(OPTION_VERSION)
-                .longOpt(OPTION_VERSION_LONG)
-                .desc("Print the PSL version and exit")
-                .build());
-
-        mainCommand.setRequired(true);
-        options.addOptionGroup(mainCommand);
-
-        options.addOption(Option.builder(OPTION_DATA)
-                .longOpt(OPTION_DATA_LONG)
-                .desc("Path to PSL data file")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder()
-                .longOpt(OPTION_DB_H2_PATH)
-                .desc("Path for H2 database file (defaults to 'cli_<user name>@<host name>' ('" + DEFAULT_H2_DB_PATH + "'))." +
-                        " Not compatible with the '--" + OPTION_DB_POSTGRESQL_NAME + "' option.")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder()
-                .longOpt(OPTION_DB_POSTGRESQL_NAME)
-                .desc("Name for the PostgreSQL database to use (defaults to " + DEFAULT_POSTGRES_DB_NAME + ")." +
-                        " Not compatible with the '--" + OPTION_DB_H2_PATH + "' option." +
-                        " Currently only local databases without credentials are supported.")
-                .hasArg()
-                .argName("name")
-                .optionalArg(true)
-                .build());
-
-        options.addOption(Option.builder(OPTION_EVAL)
-                .longOpt(OPTION_EVAL_LONG)
-                .desc("Run the named evaluator (" + Evaluator.class.getName() + ") on any open predicate with a 'truth' partition." +
-                        " If multiple evaluators are specific, they will each be run.")
-                .hasArgs()
-                .argName("evaluator ...")
-                .build());
-
-        options.addOption(Option.builder(OPTION_INT_IDS)
-                .longOpt(OPTION_INT_IDS_LONG)
-                .desc("Use integer identifiers (UniqueIntID) instead of string identifiers (UniqueStringID).")
-                .build());
-
-        options.addOption(Option.builder(OPTION_LOG4J)
-                .longOpt(OPTION_LOG4J_LONG)
-                .desc("Optional log4j properties file path")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder(OPTION_MODEL)
-                .longOpt(OPTION_MODEL_LONG)
-                .desc("Path to PSL model file")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder(OPTION_OUTPUT_DIR)
-                .longOpt(OPTION_OUTPUT_DIR_LONG)
-                .desc("Optional path for writing results to filesystem (default is STDOUT)")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder()
-                .longOpt(OPTION_OUTPUT_GROUND_RULES_LONG)
-                .desc("Output the program's ground rules." +
-                        " If a path is specified, the ground rules will be output there." +
-                        " Otherwise, they will be output to stdout (not the logger).")
-                .hasArg()
-                .argName("path")
-                .optionalArg(true)
-                .build());
-
-        options.addOption(Option.builder()
-                .longOpt(OPTION_OUTPUT_SATISFACTION_LONG)
-                .desc("Output the program's ground rules along with their satisfaction values after inference." +
-                        " If a path is specified, the ground rules will be output there." +
-                        " Otherwise, they will be output to stdout (not the logger).")
-                .hasArg()
-                .argName("path")
-                .optionalArg(true)
-                .build());
-
-        options.addOption(Option.builder(OPTION_PROPERTIES_FILE)
-                .longOpt(OPTION_PROPERTIES_FILE_LONG)
-                .desc("Optional PSL properties file path")
-                .hasArg()
-                .argName("path")
-                .build());
-
-        options.addOption(Option.builder(OPTION_PROPERTIES)
-                .argName("name=value")
-                .desc("Directly specify PSL properties (overrides options set via --" + OPTION_PROPERTIES_FILE_LONG + ")." +
-                        " See https://github.com/linqs/psl/wiki/Configuration-Options for a list of available options." +
-                        " Log4j properties (properties starting with 'log4j') will be passed to the logger." +
-                        " 'log4j.threshold=DEBUG', for example, will be passed to log4j and set the global logging threshold.")
-                .hasArg()
-                .numberOfArgs(2)
-                .valueSeparator('=')
-                .build());
-
-        return options;
-    }
-
-    private static HelpFormatter getHelpFormatter() {
-        HelpFormatter helpFormatter = new HelpFormatter();
-
-        // Hack the option ordering to put argumentions without options first and then required options first.
-        // infer and learn go first, then required, then just normal.
-        helpFormatter.setOptionComparator(new Comparator<Option>() {
-            @Override
-            public int compare(Option o1, Option o2) {
-                String name1 = o1.getOpt();
-                if (name1 == null) {
-                    name1 = o1.getLongOpt();
-                }
-
-                String name2 = o2.getOpt();
-                if (name2 == null) {
-                    name2 = o2.getLongOpt();
-                }
-
-                if (name1.equals(OPERATION_INFER)) {
-                    return -1;
-                }
-
-                if (name2.equals(OPERATION_INFER)) {
-                    return 1;
-                }
-
-                if (name1.equals(OPERATION_LEARN)) {
-                    return -1;
-                }
-
-                if (name2.equals(OPERATION_LEARN)) {
-                    return 1;
-                }
-
-                if (o1.isRequired() && !o2.isRequired()) {
-                    return -1;
-                }
-
-                if (!o1.isRequired() && o2.isRequired()) {
-                    return 1;
-                }
-
-                return name1.compareTo(name2);
-            }
-        });
-
-        helpFormatter.setWidth(100);
-
-        return helpFormatter;
-    }
-
-    /**
-     * Parse the options on the command line.
-     * Will exit on error, but Will return null if the CLI should not be run (like if we are doing a help/version run).
-     */
-    private static CommandLine parseOptions(String[] args) {
-        Options options = setupOptions();
-        CommandLineParser parser = new DefaultParser();
-        CommandLine commandLineOptions = null;
-
-        try {
-            commandLineOptions = parser.parse(options, args);
-        } catch (ParseException ex) {
-            System.err.println("Command line error: " + ex.getMessage());
-            getHelpFormatter().printHelp("psl", options, true);
-            System.exit(1);
-        }
-
-        if (commandLineOptions.hasOption(OPTION_HELP)) {
-            initDefaultLogger();
-            getHelpFormatter().printHelp("psl", options, true);
-            return null;
-        }
-
-        if (commandLineOptions.hasOption(OPTION_VERSION)) {
-            initDefaultLogger();
-            System.out.println("PSL CLI Version " + Version.getFull());
-            return null;
-        }
-
-        // Data and model are required.
-        // (We don't enforce them earlier so we can have successful runs with help and version.)
-
-        if (!commandLineOptions.hasOption(OPTION_DATA)) {
-            System.out.println(String.format("Missing required option: --%s/-%s.", OPTION_DATA_LONG, OPTION_DATA));
-            getHelpFormatter().printHelp("psl", options, true);
-            System.exit(1);
-        }
-
-        if (!commandLineOptions.hasOption(OPTION_MODEL)) {
-            System.out.println(String.format("Missing required option: --%s/-%s.", OPTION_MODEL_LONG, OPTION_MODEL));
-            getHelpFormatter().printHelp("psl", options, true);
-            System.exit(1);
-        }
-
-        // Can't have both an H2 and Postgres database.
-        if (commandLineOptions.hasOption(OPTION_DB_H2_PATH) && commandLineOptions.hasOption(OPTION_DB_POSTGRESQL_NAME)) {
-            System.err.println("Command line error: Options '--" + OPTION_DB_H2_PATH + "' and '--" + OPTION_DB_POSTGRESQL_NAME + "' are not compatible.");
-            getHelpFormatter().printHelp("psl", options, true);
-            System.exit(2);
-        }
-
-        return commandLineOptions;
-    }
-
     public static void main(String[] args) {
         main(args, false);
     }
 
     public static void main(String[] args, boolean rethrow) {
         try {
-            CommandLine commandLineOptions = parseOptions(args);
-            if (commandLineOptions == null) {
+            CommandLineLoader commandLineLoader = new CommandLineLoader(args);
+            if (!(isInputSanitized(commandLineLoader))) {
                 return;
             }
-
-            Launcher pslLauncher = new Launcher(commandLineOptions);
+            Launcher pslLauncher = new Launcher(commandLineLoader);
             pslLauncher.run();
         } catch (Exception ex) {
             if (rethrow) {
